# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Babeonym API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "schemas": {
...
      "NotAuthenticatedResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Not authenticated"
          }
        }
      },
...
      "UserSettings": {
        "type": "object",
        "required": [
          "userId",
          "theme",
          "surName"
        ],
        "properties": {
          "userId": {
            "type": "number",
            "example": 123
          },
          "theme": {
            "type": "string",
            "description": "Theme enum key",
            "example": "LIGHT"
          },
          "surName": {
            "type": "string",
            "example": "Stevens"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/user/settings": {
      "put": {
        "operationId": "v1UserSettings",
        "summary": "Update user settings",
        "description": "Updates the authenticated user's settings such as theme and surname.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "theme",
                  "surName"
                ],
                "properties": {
                  "theme": {
                    "type": "string",
                    "description": "Theme identifier"
                  },
                  "surName": {
                    "type": "string",
                    "description": "User surname"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User settings updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "settings"
                  ],
                  "properties": {
                    "settings": {
                      "$ref": "#/components/schemas/UserSettings"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (missing fields, invalid theme, or inappropriate language)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAuthenticatedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to update user settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "An error occurred while updating settings"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}